Adds support for :snapshots_into => "name"
Each step then records a snapshot, named in sequence, and for the name
of the step. There's some speed impact related to this - a 6.5 second
story goes to 8 seconds - not sure if that's additive or a factor.

I've left :snapshots_into on creates_and_edits_menu_items as a
reference, but for day to day we may want to leave it out